# sdk/openapi.json
{
  "components": {
    "schemas": {
      "Error": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "type": "object"
          }
        },
        "type": "object"
      },
...
      "SnapshotCreated": {
        "properties": {
          "snapshot_id": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "Template": {
        "properties": {
          "created_at": {
            "type": "string"
          },
          "disk_mb": {
            "type": "integer"
          },
          "egress_allow": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "error": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "image_digest": {
            "type": "string"
          },
          "memory_mb": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "sandboxes": {
            "type": "integer"
          },
          "snapshot_bytes": {
            "type": "integer"
          },
          "state": {
            "type": "string"
          },
          "ttl_seconds": {
            "type": "integer"
          },
          "vcpus": {
            "type": "integer"
          }
        },
        "type": "object"
      },
      "TemplateList": {
        "items": {
          "$ref": "#/components/schemas/Template"
        },
        "type": "array"
      },
      "TemplateState": {
        "properties": {
          "name": {
            "type": "string"
          },
          "state": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "description": "A tenant API key, or an OAuth access token.",
        "scheme": "bearer",
        "type": "http"
      },
      "oauth2": {
        "description": "An OAuth 2.1 access token of this tenant.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Isolated Linux microVMs on one host, created from OCI images and forked into copies.",
    "title": "Kiln",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/sandboxes/{id}/snapshot": {
      "post": {
        "operationId": "snapshot",
        "parameters": [
          {
            "description": "Sandbox id.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "stop": {
                    "type": "boolean"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotCreated"
                }
              }
            },
            "description": "The call succeeded."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request is invalid. The message names the field."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The credential is missing, expired or revoked."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The row does not exist, or it belongs to another tenant."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The call conflicts with the state, for example a template that still has sandboxes."
          },
          "507": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A cap of this tenant is reached. The message names the cap."
          }
        },
        "summary": "Snapshot a sandbox. With stop the sandbox sleeps and keeps its hostnames."
      }
    },
...
    "/v1/templates": {
      "get": {
        "operationId": "listTemplates",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateList"
                }
              }
            },
            "description": "The call succeeded."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request is invalid. The message names the field."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The credential is missing, expired or revoked."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The row does not exist, or it belongs to another tenant."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The call conflicts with the state, for example a template that still has sandboxes."
          },
          "507": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A cap of this tenant is reached. The message names the cap."
          }
        },
        "summary": "List the templates of this tenant."
      },
      "post": {
        "operationId": "createTemplate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "disk_mb": {
                    "type": "integer"
                  },
                  "egress_allow": {
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "image": {
                    "type": "string"
                  },
                  "memory_mb": {
                    "type": "integer"
                  },
                  "name": {
                    "type": "string"
                  },
                  "setup": {
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "ttl_seconds": {
                    "type": "integer"
                  },
                  "vcpus": {
                    "type": "integer"
                  }
                },
                "required": [
                  "name",
                  "image",
                  "egress_allow"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateState"
                }
              }
            },
            "description": "The call succeeded."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request is invalid. The message names the field."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The credential is missing, expired or revoked."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The row does not exist, or it belongs to another tenant."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The call conflicts with the state, for example a template that still has sandboxes."
          },
          "507": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A cap of this tenant is reached. The message names the cap."
          }
        },
        "summary": "Build a template from an OCI image. The build runs after the answer."
      }
    },
    "/v1/templates/{name}": {
      "delete": {
        "operationId": "deleteTemplate",
        "parameters": [
          {
            "description": "Name of the template.",
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The call succeeded."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request is invalid. The message names the field."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The credential is missing, expired or revoked."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The row does not exist, or it belongs to another tenant."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The call conflicts with the state, for example a template that still has sandboxes."
          },
          "507": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A cap of this tenant is reached. The message names the cap."
          }
        },
        "summary": "Delete a template that has no live sandbox and no snapshot."
      },
      "get": {
        "operationId": "getTemplate",
        "parameters": [
          {
            "description": "Name of the template.",
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Template"
                }
              }
            },
            "description": "The call succeeded."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request is invalid. The message names the field."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The credential is missing, expired or revoked."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The row does not exist, or it belongs to another tenant."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The call conflicts with the state, for example a template that still has sandboxes."
          },
          "507": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A cap of this tenant is reached. The message names the cap."
          }
        },
        "summary": "Read one template."
      }
    },
...
  },
  "security": [
    {
      "apiKey": []
    },
    {
      "oauth2": []
    }
  ],
  "servers": [
    {
      "url": "https://api.example.com"
    }
  ]
}